requirements.txt
chainlit
openai
langchain
pandas
numpy
seaborn
matplotlib
Pillow
google-generativeai
tabulate